refactor(domains): split validated computation portfolios
