+++ src/jacobian/domains/analysis/__init__.py
@@ -1,0 +1,5 @@
+"""Validated real-analysis capabilities."""
+
+from jacobian.domains.analysis.bundle import REAL_ANALYSIS_BUNDLE
+
+__all__ = ["REAL_ANALYSIS_BUNDLE"]
