+++ src/jacobian/domains/optimization/__init__.py
@@ -1,0 +1,5 @@
+"""Exact optimization capabilities."""
+
+from jacobian.domains.optimization.bundle import RATIONAL_OPTIMIZATION_BUNDLE
+
+__all__ = ["RATIONAL_OPTIMIZATION_BUNDLE"]
