+++ src/jacobian/domains/probability/__init__.py
@@ -1,0 +1,5 @@
+"""Exact finite-probability capabilities."""
+
+from jacobian.domains.probability.bundle import FINITE_PROBABILITY_BUNDLE
+
+__all__ = ["FINITE_PROBABILITY_BUNDLE"]
