--- src/jacobian/domains/validated_analysis/__init__.py
@@ -1,5 +1,0 @@
-"""Validated-analysis domain bundle."""
-
-from jacobian.domains.validated_analysis.bundle import VALIDATED_ANALYSIS_BUNDLE
-
-__all__ = ["VALIDATED_ANALYSIS_BUNDLE"]
